+++ views/homepage.handlebars
@@ -19,7 +19,7 @@
                 Approval Source :<br>
                 <a href="{{med.approval_source}}">extranet.who.int</a></p>
               <p>
-                <a class="mark" href="/med/{{med.id}}"><p> Comments
+                <a class="mark" href="api/med/comments/{{med.id}}"><p> Comments
                 {{#if med.comments}} 
                    : {{ med.comments.length}} </p>
                 {{/if}}  </a>
